emacs: CEDET update
Since CEDET is now included in GNU Emacs, try to use that instead of a separate
package. However this included version is quite different and most docs
available on the CEDET website seem to be outdated. This is an *attempt* to fix
things, i.e. Emacs starts correctly but the behaviour inside a real buffer has
not (yet) been tested.

Useful resource for switching from external to included CEDET:
http://stackoverflow.com/questions/3445248/configuring-cedet-for-gnu-emacs-23-2-1